Added type option to SetSelection Rev.1
